changed the wrappers to close gracefully and made the server able to connect to other servers

+++ protobufs/protobuf_compiler.py
@@ -1,11 +1,12 @@
-import os
 import shutil
 from pathlib import Path
 
 from grpc_tools import protoc
 
-COMPILE_DIR = 'compiled_protobufs'
-SOURCE_DIR = 'source_protobufs'
+BASE_DIR = Path(__file__).parent.resolve()
+
+COMPILE_DIR = BASE_DIR / "compiled_protobufs"
+SOURCE_DIR = BASE_DIR / "source_protobufs"
 
 
 def compile_protobuf(protobuf_name: str):
@@ -14,27 +15,24 @@ def compile_protobuf(protobuf_name: str):
         f"-I{SOURCE_DIR}",
         f"--python_out={COMPILE_DIR}",
         f"--pyi_out={COMPILE_DIR}",
-        protobuf_name,
+        str(SOURCE_DIR / protobuf_name),
     ])
 
 
 def recompile_all():
     clear_all()
-    src_dir = Path(SOURCE_DIR)
 
-    for file in src_dir.iterdir():
+    for file in SOURCE_DIR.iterdir():
         if file.is_file():
             compile_protobuf(file.name)
 
-    print('recompiled protobufs')
+    print("recompiled protobufs")
 
 
 def clear_all():
-    try:
+    if COMPILE_DIR.exists():
         shutil.rmtree(COMPILE_DIR)
-        os.mkdir(COMPILE_DIR)
-    except:
-        None
+    COMPILE_DIR.mkdir(parents=True, exist_ok=True)
 
 
 if __name__ == "__main__":
